backend/requirements.txt
tornado>=6.3.2
pyserial>=3.5
psutil>=5.9.0
numpy>=1.24.0
aiofiles>=23.1.0
python-periphery>=2.3.0  # For hardware GPIO access